Add Python binding for get_transitions_to_state()
